requirements.txt
anthropic>=0.40.0
supabase>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
pytz>=2024.1
